requirements.txt
annotated-types==0.7.0 ; python_version >= "3.11"
anyio==4.13.0 ; python_version >= "3.11"
attrs==26.1.0 ; python_version >= "3.11"
automat==25.4.16 ; python_version >= "3.11"
certifi==2026.2.25 ; python_version >= "3.11"
cffi==2.0.0 ; python_version >= "3.11" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.7 ; python_version >= "3.11"
colorama==0.4.6 ; python_version >= "3.11" and platform_system == "Windows"
constantly==23.10.4 ; python_version >= "3.11"
cryptography==46.0.6 ; python_version >= "3.11"
cssselect==1.4.0 ; python_version >= "3.11"
defusedxml==0.7.1 ; python_version >= "3.11"
distro==1.9.0 ; python_version >= "3.11"
et-xmlfile==2.0.0 ; python_version >= "3.11"
filelock==3.25.2 ; python_version >= "3.11"
greenlet==3.3.2 ; python_version >= "3.11"
h11==0.16.0 ; python_version >= "3.11"
httpcore==1.0.9 ; python_version >= "3.11"
httpx==0.28.1 ; python_version >= "3.11"
hyperlink==21.0.0 ; python_version >= "3.11"
idna==3.11 ; python_version >= "3.11"
incremental==24.11.0 ; python_version >= "3.11"
itemadapter==0.13.1 ; python_version >= "3.11"
itemloaders==1.4.0 ; python_version >= "3.11"
jiter==0.13.0 ; python_version >= "3.11"
jmespath==1.1.0 ; python_version >= "3.11"
lxml==6.0.2 ; python_version >= "3.11"
numpy==2.4.4 ; python_version >= "3.11"
openai==2.30.0 ; python_version >= "3.11"
openpyxl==3.1.5 ; python_version >= "3.11"
packaging==26.0 ; python_version >= "3.11"
pandas==3.0.2 ; python_version >= "3.11"
parsel==1.11.0 ; python_version >= "3.11"
playwright-stealth==1.0.6 ; python_version >= "3.11"
playwright==1.58.0 ; python_version >= "3.11"
protego==0.6.0 ; python_version >= "3.11"
pyasn1-modules==0.4.2 ; python_version >= "3.11"
pyasn1==0.6.3 ; python_version >= "3.11"
pycparser==3.0 ; implementation_name != "PyPy" and platform_python_implementation != "PyPy" and python_version >= "3.11"
pydantic-core==2.41.5 ; python_version >= "3.11"
pydantic==2.12.5 ; python_version >= "3.11"
pydispatcher==2.0.7 ; python_version >= "3.11" and platform_python_implementation == "CPython"
pyee==13.0.1 ; python_version >= "3.11"
pyopenssl==26.0.0 ; python_version >= "3.11"
pypydispatcher==2.1.2 ; python_version >= "3.11" and platform_python_implementation == "PyPy"
python-dateutil==2.9.0.post0 ; python_version >= "3.11"
python-dotenv==1.2.2 ; python_version >= "3.11"
queuelib==1.9.0 ; python_version >= "3.11"
requests-file==3.0.1 ; python_version >= "3.11"
requests==2.33.1 ; python_version >= "3.11"
schedule==1.2.2 ; python_version >= "3.11"
scrapy-playwright==0.0.46 ; python_version >= "3.11"
scrapy==2.14.2 ; python_version >= "3.11"
service-identity==24.2.0 ; python_version >= "3.11"
setuptools==70.0.0 ; python_version >= "3.11"
six==1.17.0 ; python_version >= "3.11"
sniffio==1.3.1 ; python_version >= "3.11"
tldextract==5.3.1 ; python_version >= "3.11"
tqdm==4.67.3 ; python_version >= "3.11"
twisted==25.5.0 ; python_version >= "3.11"
typing-extensions==4.15.0 ; python_version >= "3.11"
typing-inspection==0.4.2 ; python_version >= "3.11"
tzdata==2026.1 ; python_version >= "3.11" and (sys_platform == "win32" or sys_platform == "emscripten")
urllib3==2.6.3 ; python_version >= "3.11"
w3lib==2.4.1 ; python_version >= "3.11"
zope-interface==8.2 ; python_version >= "3.11"
